back-end/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
passlib[bcrypt]==1.7.4
python-decouple==3.8
Pillow==11.0.0
email-validator==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
